Reduced the strength of bone armor because it made a full outfit near invulnerable. Enabled Anu achievement Added new Compsognathus sounds Modified TRex model Fixed a method not being called on all dinosaurs regarding receiving attack damage TRex and Spinosaur cannot be hurt by bare hands Fixed being able to feed wild mosasaur Dinopedia now works with Nautilus Sifters now drop biofossil instead of DNA. Anu can be placed through spawn egg Changed dilophosaurus hurt sound

+++ src/main/java/mods/fossil/entity/mob/EntityDilophosaurus.java
@@ -364,7 +364,7 @@ public class EntityDilophosaurus extends EntityDinosaur
     /**
      * Called when the entity is attacked.
      */
-    public boolean attackEntityFrom(DamageSource var1, int var2)
+    public boolean attackEntityFrom(DamageSource var1, float var2)
     {
         Entity var3 = var1.getEntity();
         boolean var4 = false;
