+++ src/main/java/mods/fossil/entity/mob/EntityStegosaurus.java
@@ -213,7 +213,7 @@ public class EntityStegosaurus extends EntityDinosaur
     /**
      * Called when the entity is attacked.
      */
-    public boolean attackEntityFrom(DamageSource var1, int var2)
+    public boolean attackEntityFrom(DamageSource var1, float var2)
     {
         Entity var3 = var1.getEntity();
         this.setSitting(false);
